# docs/swagger.yaml
basePath: /
definitions:
  analitik.DTOHistogramDefect:
    properties:
      jumlah:
        description: Frekuensi/jumlah defect pada kategori tersebut
        example: 15
        type: number
      kategori:
        description: 'Kategori grouping (contoh: slot waktu, tanggal, kode)'
        example: 08:00-12:00
        type: string
    type: object
  analitik.DTORingkasanNG:
    properties:
      jumlah_hari:
        description: Jumlah hari aktif produksi
        example: 5
        type: integer
      rasio_ng:
        description: Persentase NG (%)
        example: 5
        type: number
      total_defect:
        description: Total barang NG (Reject)
        example: 50
        type: number
      total_ok:
        description: Total barang OK (Lolos QC)
        example: 950
        type: number
      total_produksi:
        description: Total barang diproduksi
        example: 1000
        type: number
    type: object
  analitik.DTOSinyalKualitas:
    properties:
      alasan:
        description: Penjelasan status
        example: Rasio NG di bawah 2%
        type: string
      indikator:
        description: 'Metrik utama yang diukur (contoh: rasio NG)'
        example: 1.5
        type: number
      status:
        description: KRITIS, WASPADA, atau STABIL
        example: STABIL
        type: string
    type: object
  analitik.DTOStratifikasiDefect:
    properties:
      dimensi:
        description: Dasar stratifikasi
        example: kode_cacat
        type: string
      jumlah:
        description: Jumlah defect
        example: 35
        type: number
      kategori:
        description: Nilai kategori stratifikasi
        example: A
        type: string
    type: object
  analitik.DTOTrendDefect:
    properties:
      jumlah_defect:
        description: Frekuensi defect
        example: 12
        type: number
      periode:
        description: Waktu periode (harian, mingguan, bulanan)
        example: "2026-05-01"
        type: string
    type: object
...
  respon.DetailKesalahan:
    properties:
      field:
        type: string
      pesan:
        type: string
    type: object
  respon.ResponStandar:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      kesalahan:
        items:
          $ref: '#/definitions/respon.DetailKesalahan'
        type: array
      message:
        type: string
      meta: {}
      metadata: {}
      pesan:
        type: string
      status:
        type: string
      success:
        description: Legacy fields untuk retro-kompatibilitas
        type: boolean
      sukses:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API untuk ekosistem manufaktur dan kontrol kualitas PGNServer.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PGNServer API
  version: "1.0"
paths:
  /api/v1/analitik/histogram_defect:
    get:
      consumes:
      - application/json
      description: Mengembalikan frekuensi defect yang dikelompokkan berdasarkan parameter
        tertentu.
      parameters:
      - description: Tanggal Mulai (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Tanggal Selesai (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Nama Lini / Zona
        in: query
        name: line
        type: string
      - description: Grup (waktu | tanggal | kode_cacat)
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/analitik.DTOHistogramDefect'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Dapatkan Histogram Defect
      tags:
      - Analitik
  /api/v1/analitik/lacak:
    get:
      consumes:
      - application/json
      description: Menelusuri defects produk hingga ke level bahan baku pembentuk
        & supplier
      parameters:
      - description: Kode Defect Cacat
        in: query
        name: kode_cacat
        required: true
        type: string
      - description: SKU Produk Jadi (Finished Good)
        in: query
        name: parent_sku
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Lacak Akar Masalah Defect (BOM Tracing)
      tags:
      - Analitik
  /api/v1/analitik/metrik_pareto_bulanan:
    get:
      consumes:
      - application/json
      description: Mengembalikan kalkulasi Pareto 80/20 per bulan menggunakan Window
        Function SQL
      parameters:
      - description: Bulan (1-12)
        in: query
        name: bulan
        type: integer
      - description: 'Tahun (contoh: 2026)'
        in: query
        name: tahun
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Dapatkan Metrik Pareto Bulanan
      tags:
      - Analitik
  /api/v1/analitik/pareto:
    get:
      consumes:
      - application/json
      description: Mengembalikan kalkulasi Pareto 80/20 berdasarkan tanggal mulai,
        tanggal selesai, dan lini.
      parameters:
      - description: Tanggal Mulai (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Tanggal Selesai (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Nama Lini / Zona
        in: query
        name: line
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Dapatkan Pareto Dinamis
      tags:
      - Analitik
  /api/v1/analitik/ringkasan_ng:
    get:
      consumes:
      - application/json
      description: Mengembalikan ringkasan produksi dan defect (Total Produksi, OK,
        Defect, Rasio NG).
      parameters:
      - description: Tanggal Mulai (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Tanggal Selesai (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Nama Lini / Zona
        in: query
        name: line
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  $ref: '#/definitions/analitik.DTORingkasanNG'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Dapatkan Ringkasan NG
      tags:
      - Analitik
  /api/v1/analitik/sinyal_kualitas:
    get:
      consumes:
      - application/json
      description: Mengembalikan sinyal peringatan kualitas (KRITIS, WASPADA, STABIL)
        berdasarkan batas toleransi.
      parameters:
      - description: Tanggal Mulai (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Tanggal Selesai (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Nama Lini / Zona
        in: query
        name: line
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  $ref: '#/definitions/analitik.DTOSinyalKualitas'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Dapatkan Sinyal Kualitas
      tags:
      - Analitik
  /api/v1/analitik/stratifikasi_defect:
    get:
      consumes:
      - application/json
      description: 'Mengembalikan stratifikasi defect (contoh: berdasarkan kode_cacat).'
      parameters:
      - description: Tanggal Mulai (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Tanggal Selesai (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Nama Lini / Zona
        in: query
        name: line
        type: string
      - description: Kode Cacat Spesifik
        in: query
        name: kode_cacat
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/analitik.DTOStratifikasiDefect'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Dapatkan Stratifikasi Defect
      tags:
      - Analitik
  /api/v1/analitik/trend_defect:
    get:
      consumes:
      - application/json
      description: Mengembalikan tren defect berdasarkan periode.
      parameters:
      - description: Tanggal Mulai (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Tanggal Selesai (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Nama Lini / Zona
        in: query
        name: line
        type: string
      - description: Periode (harian | mingguan | bulanan)
        in: query
        name: periode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/respon.ResponStandar'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/analitik.DTOTrendDefect'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Dapatkan Trend Defect
      tags:
      - Analitik
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"